Refactor to be called by  webhooks
We were reaching our integration limit whilst on the free account. So
we no longer run this as a slack bot (which counts to the
integration limit) and instead configure it to run as a webhook
responding to certain words and phrases.

This has the added security benefit that only these slack messages will
be sent to the bot instead of all traffic.

A secondary benefit is that the list of which github teams or repos to
query for open PRs can be part of the message - meaning that this app
will not require a datastore indicating which channels should poll which
repos.