Use GitHub GraphQL API instead of REST
[GitHub's GraphQL API](https://developer.github.com/v4/) is
significantly faster for the sort of heavily linked queries we're doing
because they handle it all in one request.

This completely changes the method by which we request the data to use
the GraphQL API.

For the moment, I've also removed the ability to list who issues are
assigned to, because that involves walking further down the graph, which
just complicates matters more than we really need.