Show pull requests' age in output
It's often useful to get context on how old the pull requests are, so
we can know what's no longer relevant (and in need of closing) or has
been open longer than we'd like (and needs attention).

This is a bit hacky in terms of the maths, but I think it's right.