Add a deprecation notice, suggesting Pull Panda
This bot exhausted the Heroku free usage tier, this month.
Rather than pay to host it, which is another thing we have to
manage on an ongoing basis, this commit suggests that people use
Pull Panda instead, which offers a superset of the functionality
for free (since GitHub acquired Pull Panda).